moved client's code to default directories
web client's code removed to  default  directories
