+++ src/main/client/js/components/ProjectsList/ProjectsList.scss
