+++ src/main/client/styles/base/variables.scss
