+++ src/main/client/styles/components/layout.scss
