+++ src/main/client/styles/components/new-moon.scss
