+++ src/main/client/styles/components/scaffolding.scss
