+++ src/main/client/styles/components/tables.scss
