+++ src/main/java/com/eugenem/blog/ReactAndSpringDataRestApplication.java
@@ -1,0 +1,16 @@
+package com.eugenem.blog;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * @author eugenem
+ */
+
+@SpringBootApplication
+public class ReactAndSpringDataRestApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ReactAndSpringDataRestApplication.class, args);
+	}
+}
