+++ src/main/resources/static/favicon/favicon.ico
