+++ src/main/resources/templates/index.html
@@ -25,6 +25,7 @@
         ga('send', 'pageview');
     </script>
     <!-- End Google Analytics -->
+    <link href="./style.css" rel="stylesheet">
 </head>
 <body>
     <div class="header"></div>
@@ -33,6 +34,6 @@
     </div>
     <div class="footer">
     </div>
-    <script src="/bundle.js"></script>
+    <script src="bundle.js"></script>
 </body>
 </html>
