+++ src/main/resources/templates/info.html
@@ -1,0 +1,6 @@
+<html>
+    <head></head>
+    <body>
+        <h2>Information</h2>
+    </body>
+</html>
